* rename app -> trainer_app * add `setup.py`
