+++ trainer.py
@@ -1,4 +1,4 @@
-from app import create_app, cli
+from trainer_app import create_app, cli
 
 app = create_app()
 cli.register(app)
