+++ trainer_app/cli.py
@@ -12,7 +12,7 @@ def register(app):
     @utils.command()
     def test():
         """Runs the unit tests."""
-        tests = unittest.TestLoader().discover('app/test', pattern='test*.py')
+        tests = unittest.TestLoader().discover('trainer_app/test', pattern='test*.py')
         result = unittest.TextTestRunner(verbosity=2).run(tests)
         if result.wasSuccessful():
             return 0
