+++ trainer_app/front/templates/index.html
