+++ trainer_app/front/templates/layout.html
