+++ trainer_app/rest/models/dto.py
