+++ trainer_app/trainer/main_query.sql
