+++ trainer_app/trainer/utils.py
